docs refactor
