+++ doxygen/docs.py
@@ -41,7 +41,8 @@ REGEX_COMMENT = re.compile(r'\s*// ?(.*)\n?$')
 assert REGEX_COMMENT.match("   // AUI_DOCS_CODE_BEGIN\n")
 
 # TEST_F(UIDataBindingTest, AProperty) { // HEADER
-REGEX_TESTCASE_HEADER = re.compile(r'TEST_F\(.+, (.+)\) ?\{ ?// ?HEADER')
+REGEX_TESTCASE_HEADER = re.compile(r'TEST_F\((.+), (.+)\) ?\{ ?// ?HEADER')
+REGEX_TESTCASE_HEADER_H1 = re.compile(r'TEST_F\((.+), (.+)\) ?\{ ?// ?HEADER_H1')
 
 REGEX_INGROUP = re.compile(r'.*([@\\]ingroup ?\w*).*')
 assert REGEX_INGROUP.match("   * @ingroup")
@@ -129,12 +130,22 @@ def process_cpp_file(input: Path):
                 CODE_END = "@endcode"
 
             for line in lines:
+                if match := REGEX_TESTCASE_HEADER_H1.match(line):
+                    emit_line()
+                    fos.write("# ")
+                    fos.write(match.group(2).replace("_", " "))
+                    fos.write(" {#")
+                    fos.write(f'{match.group(1)}_{match.group(2)}')
+                    fos.write("}")
+                    emit_line()
+                    continue
+
                 if match := REGEX_TESTCASE_HEADER.match(line):
                     emit_line()
                     fos.write("## ")
-                    fos.write(match.group(1).replace("_", " "))
+                    fos.write(match.group(2).replace("_", " "))
                     fos.write(" {#")
-                    fos.write(match.group(1))
+                    fos.write(f'{match.group(1)}_{match.group(2)}')
                     fos.write("}")
                     emit_line()
                     continue
